+++ backend/src/users/users.module.ts
@@ -4,8 +4,8 @@ import { UsersService } from './users.service';
 import { User } from './user.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([User])],
-  providers: [UsersService],
-  exports: [UsersService],
+    imports: [TypeOrmModule.forFeature([User])],
+    providers: [UsersService],
+    exports: [UsersService],
 })
-export class UsersModule {}
+export class UsersModule { }